LayoutTests/css3/calc/line-height-expected.txt

Tests that CSS3 calc() can be used for the line-height property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getComputedStyle(document.getElementById("calc-percent"), null).lineHeight should be 32px. Was normal.
FAIL getComputedStyle(document.getElementById("calc-percent-pixels"), null).lineHeight should be 32px. Was normal.
PASS successfullyParsed is true

TEST COMPLETE
